% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_document_id.R
\name{add_document_id}
\alias{add_document_id}
\title{`add_document_id` adds ID column to document_dataframe}
\usage{
add_document_id(dataframe)
}
\arguments{
\item{dataframe}{}
}
\value{
dataframe
}
\description{
Adds a column to a dataframe that contains ID's
}
