% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_url.R
\name{remove_url}
\alias{remove_url}
\title{`remove_url` removes any URLs from the text}
\usage{
remove_url(text)
}
\arguments{
\item{text}{}
}
\value{

}
\description{
`remove_url` removes any URLs from the text
}
